Rename courseDatasource into demoDatasource
