+++ api/lib/infrastructure/datasources/airtable/index.js
@@ -1,7 +1,7 @@
 const AreaDatasource = require('./area-datasource');
 const ChallengeDatasource = require('./challenge-datasource');
 const CompetenceDatasource = require('./competence-datasource');
-const CourseDatasource = require('./course-datasource');
+const DemoDatasource = require('./demo-datasource');
 const SkillDatasource = require('./skill-datasource');
 const TubeDatasource = require('./tube-datasource');
 const TutorialDatasource = require('./tutorial-datasource');
@@ -10,7 +10,7 @@ module.exports = {
   AreaDatasource,
   ChallengeDatasource,
   CompetenceDatasource,
-  CourseDatasource,
+  DemoDatasource,
   SkillDatasource,
   TubeDatasource,
   TutorialDatasource,
