+++ api/tests/unit/infrastructure/repositories/course-repository_test.js
@@ -2,7 +2,7 @@ const { expect, sinon } = require('../../../test-helper');
 
 const courseRepository = require('../../../../lib/infrastructure/repositories/course-repository');
 const Course = require('../../../../lib/domain/models/Demo');
-const courseDatasource = require('../../../../lib/infrastructure/datasources/airtable/course-datasource');
+const courseDatasource = require('../../../../lib/infrastructure/datasources/airtable/demo-datasource');
 
 describe('Unit | Repository | course-repository', function() {
